clk: altera-a10-fpll: Fix calibration for phase aligned mode
In phase aligned mode the fPLL uses an external feedback path that allows
to better align the PLL output phase to the phase of the external reference
clock signal.

This allows for example to support reliable sampling of source synchronous
interfaces.

But when the PLL is configured for external feedback path mode the PLL
calibration can fail and choose the wrong VCO band. The latest Intel
transceiver datasheet contains the proper sequence to make calibration work
in phase aligned mode.

This involves writing to an otherwise undocumented register to select the
internal feedback path while running the calibration. Once the calibration
is done restore the original feedback path setting.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>