jesd204: sysfs: show JESD204 link state in sysfs
The state attribute is private (part of the jesd204_link_opaque struct) and
we don't want to move it from there. If we do that, we expose a bit more of
the FSM data, which could allow drivers to do more hacks into the framework.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 8429201c4a52097c3aed3236f4f8c1bb63f4a065)