jesd204: adrv9009,ad9081: change semantic num_links -> max_num_links
In the initial phase of the JESD204 framework work, it wasn't all that
clear where the number of JESD204 links would be specified, so the init
data passed to the framework contained the exact number of JESD204 links
that a topology should contain.

However, with the latest code, the number of "active" links is specified in
the device-tree. So, for this 'num_links' on the init-data we can
re-purpose it to be 'max_num_links' to make sure the DT doesn't go above
the number of JESD204 links that a device can support.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit ce4604217d523e81d168d788c026188cbd2959ee)